Use utility to derive header user
